+++ src/Controller/Admin/DashboardController.php
@@ -5,6 +5,7 @@ namespace App\Controller\Admin;
 use App\Entity\DataSource;
 use App\Entity\ProcessOverview;
 use App\Entity\ProcessOverviewGroup;
+use App\Entity\UserRole;
 use EasyCorp\Bundle\EasyAdminBundle\Attribute\AdminDashboard;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Dashboard;
 use EasyCorp\Bundle\EasyAdminBundle\Config\MenuItem;
@@ -42,9 +43,12 @@ class DashboardController extends AbstractDashboardController
 
     public function configureMenuItems(): iterable
     {
-        yield MenuItem::linkToCrud(t('Group'), null, ProcessOverviewGroup::class);
-        yield MenuItem::linkToCrud(t('Process overview'), null, ProcessOverview::class);
-        yield MenuItem::linkToCrud(t('Data source'), null, DataSource::class);
+        yield MenuItem::linkToCrud(t('Group'), null, ProcessOverviewGroup::class)
+            ->setPermission(UserRole::OverviewEditor->value);
+        yield MenuItem::linkToCrud(t('Process overview'), null, ProcessOverview::class)
+            ->setPermission(UserRole::OverviewEditor->value);
+        yield MenuItem::linkToCrud(t('Data source'), null, DataSource::class)
+            ->setPermission(UserRole::SuperAdmin->value);
 
         yield MenuItem::section();
         yield MenuItem::linkToRoute(t('Home'), null, 'app_default');