+++ src/Controller/Admin/ProcessOverviewGroupCrudController.php
@@ -3,15 +3,18 @@
 namespace App\Controller\Admin;
 
 use App\Entity\ProcessOverviewGroup;
+use App\Entity\UserRole;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Action;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Actions;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Field\DateTimeField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\IdField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 use function Symfony\Component\Translation\t;
 
+#[IsGranted(UserRole::OverviewEditor->value)]
 class ProcessOverviewGroupCrudController extends AbstractCrudController
 {
     public static function getEntityFqcn(): string