+++ src/Entity/UserRole.php
@@ -6,5 +6,8 @@ enum UserRole: string
 {
     case User = 'ROLE_USER';
     case Admin = 'ROLE_ADMIN';
-    case OverviewManager = 'ROLE_OVERVIEW_MANAGER';
+    case SuperAdmin = 'ROLE_SUPER_ADMIN';
+    case OverviewEditor = 'ROLE_OVERVIEW_EDITOR';
+    case OverviewViewer = 'ROLE_OVERVIEW_VIEWER';
+    case ProcessStepRunner = 'ROLE_PROCESS_STEP_RUNNER';
 }