+++ src/ProcessOverviewHelper.php
@@ -3,10 +3,12 @@
 namespace App;
 
 use App\Entity\ProcessOverview;
+use App\Entity\UserRole;
 use League\Uri\Modifier;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Component\Yaml\Yaml;
 
 class ProcessOverviewHelper
@@ -15,6 +17,7 @@ class ProcessOverviewHelper
         private readonly PropertyAccessorInterface $propertyAccessor,
         private readonly DataSourceHelper $dataSourceHelper,
         private readonly UrlGeneratorInterface $urlGenerator,
+        private readonly AuthorizationCheckerInterface $authorizationChecker,
     ) {
     }
 
@@ -89,15 +92,21 @@ class ProcessOverviewHelper
                     ),
                     array_map(
                         function (array $step) use ($overview) {
-                            return $step + [
+                            $additionalStepData = [
                                 'type' => 'step',
-                                'rerun_url' => $step['can_rerun'] ? $this->urlGenerator->generate('process_overview_rerun',
-                                    [
+                            ];
+
+                            if ($this->authorizationChecker->isGranted(UserRole::ProcessStepRunner->value)) {
+                                $additionalStepData['rerun_url'] = $step['can_rerun']
+                                    ? $this->urlGenerator->generate('process_overview_rerun', [
                                         'group' => $overview->getGroup()->getId(),
                                         'overview' => $overview->getId(),
                                         'run' => $step['id'],
-                                    ], UrlGeneratorInterface::ABSOLUTE_URL) : null,
-                            ];
+                                    ], UrlGeneratorInterface::ABSOLUTE_URL)
+                                    : null;
+                            }
+
+                            return $step + $additionalStepData;
                         },
                         $steps
                     ),